fix(md2pptx): fix file extension on image download
Currently when downloading a referenced image, if
its file extension is inferred from URL suffix,
there will be a dot in the front of suffix; there
won't be if matching Content-Type.

Add a dot in the front of `suffix` on HTTP
Response with a Content-Type header to make file
name with extension in order to insert the image
into pptx file correctly.